+++ plugins/postcss-hwb-function/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS HWB Function
 
+### Unreleased (patch)
+
+- Reuse code with `@csstools/utilities`
+
 ### 3.0.8
 
 _December 31, 2023_